Add salt and pepper noise generator
